Fix inlining of `CFString` literals
I'm not sure when this broke but on 4.3.6550-dev (1ae64a95) inlining of `CFString` literals when viewing a Mach-O binary is not working at all. This is caused by the type of global constant `CFString`s in the `__cfstring` section being typed `__NSConstantString`. However this plugin is expecting them to be `CFString`.

This PR adds an additional check on the string's variable type to see if it is an `__NSConstantString`. There's probably a more robust solution to this but it at least gets the `CFString` inlining working again.