Add a setting to opt out of rewriting message sends with visible implementations
The setting is on by default to preserve the current experience. It
should likely be revisited as related changes are merged.

I find this to be confusing as often as it is helpful since it picks the
first implementation of the selector it sees without consideration for
the type of the receiver.

This is particularly annoying if the binary being analyzed implements a
method with the same name as a commonly-used method on a system type
(`-description` or `-path`, for instance), or has methods with generic
names (`-initWithURL:`) on many types.

Explicit cross-references from selectors to method implementations make
it possible to see the potential implementations without rewriting the
call, and an Objective-C pseudo-language
(https://github.com/Vector35/binaryninja-api/pull/6807) provides an even
more natural representation of Objective-C message sends without these
downsides.